feature(ux): update download flow
