+++ routes/telegram.php
@@ -1,7 +1,9 @@
 <?php
 
+use App\Telegram\Actions\DownloadAction;
 use App\Telegram\Actions\SearchYoutubeAction;
 
 return [
+    'download' => DownloadAction::class,
     'search-youtube' => SearchYoutubeAction::class,
 ];
